Take advantage of ping stack response caching to make layout changes seamless rather than reparenting components.
The reparenting approach is a bit hacky, and happens to work but isn't a
documented use of the `render` function, so it's better to simply
re-render (quickly) when the layout changes. Layout changes are not
high-volume, so it doesn't need to be as performant as it could be.
There was also at least once where I was able to get the reparenting to
fail (not with an exception, though), and the new container stayed
empty.